requirements.txt
# =============================================================================
# ГЛАВНЫЙ ФАЙЛ ЗАВИСИМОСТЕЙ ПРОЕКТА
# Все требования инструментов собраны здесь. Установка: pip install -r requirements.txt
# =============================================================================

# --- Панель (backend + веб) ---
flask>=3.0.0
flask-cors>=4.0.0

# --- Anonymous SMS ---
requests>=2.31.0

# --- Instagram Bruter ---
beautifulsoup4>=4.12.0
colorama>=0.4.0
pysocks>=1.7.0
requests-html>=0.10.0
lxml_html_clean>=0.3.0

# --- Общие / модуль domain_search ---
dnspython>=2.4.0

# --- Holehe (поиск email по сайтам) ---
holehe>=1.61
termcolor>=2.0.0
httpx>=0.24.0
trio>=0.22.0
tqdm>=4.65.0

# --- h8mail (утечки по email) ---
h8mail

# --- Sherlock (поиск username в соцсетях) ---
sherlock-project>=0.14.0
certifi>=2019.6.16
requests-futures>=1.0.0
stem>=1.8.0
pandas>=2.0.0
openpyxl>=3.0.10
tomli>=2.0.0

# --- Maigret (поиск username по сайтам) ---
maigret>=0.5.0
aiohttp>=3.9.0
aiodns>=3.0.0
socid-extractor>=0.0.27

# --- theHarvester (домен: email, поддомены) ---
theHarvester
PyYAML>=6.0

# --- Прочее (телефоны и т.д.) ---
phonenumbers>=8.13.0

# Subfinder — Go-бинарник, в requirements не входит.
# Сборка: cd tools/subfinder && go build -o subfinder ./cmd/subfinder

# Презентация (опционально)
python-pptx>=0.6.21
